Improvement model working again, but not learning
