+++ runs/Improvement/configuration.cfg
@@ -4,8 +4,6 @@ environment = Improvement
 trainer = REINFORCE
 train_batch_size = 128
 train_batch_epoch = 10
-[environment]
-T = 3
 [baseline]
 baselineType = ExpMovingAvg
 [actor]
@@ -15,7 +13,8 @@ dim_model = 128
 dim_hidden = 512
 dim_k = 128
 dim_v = 128
-[trainer]
+[optimiser]
+T = 3
 maxGrad = 2
 actor_lr = 1e-4
 actor_lr_gamma = 1
